requirements.txt
# 核心数据处理
pandas>=1.5.0
numpy>=1.24.0
pyarrow>=12.0.0

# 回测与性能
#pandas-ta>=0.3.14b0

# 机器学习（可选）
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0

# 数据源
tushare>=1.2.89
akshare>=1.11.0

# 基础设施
redis>=4.5.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# 日志与配置
pyyaml>=6.0
python-json-logger>=2.0.0

# 监控
prometheus-client>=0.17.0

# 测试与开发
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
pre-commit>=3.0.0

python-dotenv>=0.19.0

# 可选：实盘接口（根据需求安装）
# qmt-python  # 迅投 QMT
# huatai-api  # 华泰证券
